made some progress on fsys, fixed size_t error in array and core with uint64_t

+++ C++/array/dvec.hpp
@@ -30,41 +30,43 @@
   v.print();		//prints whole matrix
   v.info();		//prints info about matrix
   v.zero();		//zeros the whole matrix
-  v.size();		//returns number of elem (size_t)
+  v.size();		//returns number of elem (uint64_t)
 
 --------------------------------------------------------*/
 #ifndef DVEC_HPP
 #define DVEC_HPP
 
-#include <cstdlib> //for size_t
+#include <cstdlib>
+#include <cstdint> //for integer types
 #include <stdio.h> //for printf
 #include <limits>  //for numeric_limits::max()
+#include <assert.h>  //for assert
 
 class dvec
 {
   private:
   double*           buf;		//pointer to start of memory
-  std::size_t       len;		//number of elements
+  std::uint64_t     len;		//number of elements
   bool        allocated;                //is allocated
   bool         assigned;                //is assigned
 
   public:
   //initialization/destructors
   dvec();					//empty constructor
-  dvec(const std::size_t n);			//construct with n elements
-  dvec(const std::size_t n, double* ptr);	//construct with n elements
+  dvec(const std::uint64_t n);			//construct with n elements
+  dvec(const std::uint64_t n, double* ptr);	//construct with n elements
  ~dvec();					//destructor 
 
   //Operator overloading
-  double& operator() (const std::size_t i);	 //ref elm i
-  double operator() (const std::size_t i) const; //const elm i
+  double& operator() (const std::uint64_t i);	 //ref elm i
+  double operator() (const std::uint64_t i) const; //const elm i
 
   //Class functions
-  void allocate(const std::size_t n);			//allocate memory 
+  void allocate(const std::uint64_t n);			//allocate memory 
   void deallocate();					//deallocate memory
-  void assign(const std::size_t n, double* ptr);	//assign to memory
+  void assign(const std::uint64_t n, double* ptr);	//assign to memory
   void unassign();					//unassign to memory
-  std::size_t size() const;				//return length
+  std::uint64_t size() const;				//return length
   void info() const;					//print info
   void print() const;					//prints the elements of the vector
   void zero();						//zeros the vector
